Добавил конфигурацию запуска приложения с заглушками
