+++ src/app/environment.ts
@@ -1,20 +1,20 @@
-import {AudienceMockService} from "./services/mock/audience-mock.service";
-import {SubjectMockService} from "./services/mock/subject-mock.service";
-import {GroupMockService} from "./services/mock/group-mock.service";
-import {TeacherMockService} from "./services/mock/teacher-mock.service";
-import {ScheduleMockService} from "./services/mock/schedule-mock.service";
-import {LessonMockService} from "./services/mock/lesson-mock.service";
-import {UserMockService} from "./services/mock/user-mock.service";
-import {RequestMockService} from "./services/mock/request-mock.service";
+import {GroupService} from "./services/dao/group.service";
+import {AudienceService} from "./services/dao/audience.service";
+import {SubjectService} from "./services/dao/subject.service";
+import {TeacherService} from "./services/dao/teacher.service";
+import {UserService} from "./services/dao/user.service";
+import {RequestService} from "./services/dao/request.service";
+import {ScheduleService} from "./services/dao/schedule.service";
+import {LessonService} from "./services/dao/lesson.service";
 
 export const environment = {
   production: false,
-  groupService: GroupMockService,
-  audienceService: AudienceMockService,
-  subjectService: SubjectMockService,
-  teacherService: TeacherMockService,
-  scheduleService: ScheduleMockService,
-  lessonService: LessonMockService,
-  userService: UserMockService,
-  requestService: RequestMockService
+  groupService: GroupService,
+  audienceService: AudienceService,
+  subjectService: SubjectService,
+  teacherService: TeacherService,
+  scheduleService: ScheduleService,
+  lessonService: LessonService,
+  userService: UserService,
+  requestService: RequestService
 };
